% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weinerProcessFunctions.R
\name{correlateWeinerProcess}
\alias{correlateWeinerProcess}
\title{Construct a correlated Weiner process}
\usage{
correlateWeinerProcess(X, Sigma, dt)
}
\arguments{
\item{X}{A matrix of real values. The iid Normal random variables.}

\item{Sigma}{A covariance matrix. The instantaneous covariance of the Weiner process.}

\item{dt}{A positive real value. The size of each time step.}
}
\description{
Map iid standard Normal random variables to a correlated Weiner process
}
