% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cirProcessFunctions.R
\name{updateProcessValue_cir}
\alias{updateProcessValue_cir}
\title{Update the value of a CIR process}
\usage{
updateProcessValue_cir(xt, dt, alpha, mu, sigma, dZt)
}
\arguments{
\item{xt}{A positive real scalar. The current value of the process.}

\item{dt}{A positive real scalar. The time step to update over.}

\item{alpha}{A positive real scalar. A parameter of the CIR process.}

\item{mu}{A positive real scalar. A parameter of the CIR process.}

\item{sigma}{A positive real scalar. A parameter of the CIR process.}

\item{dZt}{A real scalar. The realised change in value of a Weiner process.}
}
\value{
A positive real scalar. The updated value of the process.
}
\description{
Update the value of a CIR process
}
